requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
scipy
statsmodels
xgboost
biopython
ViennaRNA
python_codon_tables
webencodings
tinycss2
